Shader: Move shader code to its own subdirectory, "shader".
