+++ src/citra_qt/debugger/graphics_vertex_shader.cpp
@@ -8,7 +8,7 @@
 #include <QBoxLayout>
 #include <QTreeView>
 
-#include "video_core/shader_interpreter.h"
+#include "video_core/shader/shader_interpreter.h"
 
 #include "graphics_vertex_shader.h"
