+++ src/video_core/CMakeLists.txt
@@ -11,7 +11,7 @@ set(SRCS
             pica.cpp
             primitive_assembly.cpp
             rasterizer.cpp
-            shader_interpreter.cpp
+            shader/shader_interpreter.cpp
             utils.cpp
             video_core.cpp
             )
@@ -35,7 +35,7 @@ set(HEADERS
             primitive_assembly.h
             rasterizer.h
             renderer_base.h
-            shader_interpreter.h
+            shader/shader_interpreter.h
             utils.h
             video_core.h
             )
