+++ src/video_core/clipper.cpp
@@ -7,7 +7,7 @@
 #include "clipper.h"
 #include "pica.h"
 #include "rasterizer.h"
-#include "shader_interpreter.h"
+#include "shader/shader_interpreter.h"
 
 namespace Pica {
