+++ src/video_core/primitive_assembly.h
@@ -8,7 +8,7 @@
 
 #include "video_core/pica.h"
 
-#include "video_core/shader_interpreter.h"
+#include "video_core/shader/shader_interpreter.h"
 
 namespace Pica {
