+++ src/video_core/renderer_opengl/gl_rasterizer.h
@@ -9,7 +9,7 @@
 #include "common/common_types.h"
 
 #include "video_core/hwrasterizer_base.h"
-#include "video_core/shader_interpreter.h"
+#include "video_core/shader/shader_interpreter.h"
 
 #include "gl_state.h"
 #include "gl_rasterizer_cache.h"
